Initial commit
I just wrapped up the last lesson in Jorge Escobar's "Zero to Flask: The Professional Way" course on Skillshare. This is the state of my code at that moment. It's really just the code from the course, though I did clean up some PEP8 warnings, update some of the dependencies, and work out a few minor problems that arose from that, but it's still 99% just the codebase that's covered in that class.