Comment.get_commenter()
Returns the author's full_name value if the comment has an author, or the commenter_name value if it isn't. This allows for anonymous comments and for unauthenticated users to supply a name, but if an author makes a comment and then changes her name, that change is reflected across all of her comments as well.